backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend application for the budgetmanager"
requires-python = ">=3.8,<3.10"
dependencies = [
    "django-cors-headers==3.11.0",
    "django==4.2.1",
    "djangorestframework==3.14.0",
    "requests==2.31.0",
]

[tool.uv]
dev-dependencies = [
    "black>=24.8.0",
    "isort>=5.13.2",
    "mypy>=1.14.1",
    "python-dotenv>=1.0.1",
]
